+++ Sora/Sora.cs
@@ -60,7 +60,7 @@ public class Sora : IDisposable
         public bool? Ordered;
         public int? MaxPacketLifeTime;
         public int? MaxRetransmits;
-        public string Protocol;
+        public string? Protocol;
         public bool? Compress;
     }
 
@@ -74,7 +74,7 @@ public class Sora : IDisposable
 
     public class ForwardingFilter
     {
-        public string Action;
+        public string? Action;
         public class Rule
         {
             public string Field;
@@ -82,6 +82,8 @@ public class Sora : IDisposable
             public List<string> Values = new List<string>();
         }
         public List<List<Rule>> Rules = new List<List<Rule>>();
+        public string? Version;
+        public string? Metadata;
     }
 
     /// <summary>
@@ -457,7 +459,11 @@ public class Sora : IDisposable
         if (config.ForwardingFilter != null)
         {
             cc.enable_forwarding_filter = true;
-            cc.forwarding_filter.action = config.ForwardingFilter.Action;
+            if (config.ForwardingFilter.Action != null)
+            {
+                cc.forwarding_filter.enable_action = true;
+                cc.forwarding_filter.action = config.ForwardingFilter.Action;
+            }
             foreach (var rs in config.ForwardingFilter.Rules)
             {
                 var ccrs = new SoraConf.Internal.ForwardingFilter.Rules();
@@ -473,6 +479,16 @@ public class Sora : IDisposable
                     ccrs.rules.Add(ccr);
                 }
                 cc.forwarding_filter.rules.Add(ccrs);
+            }
+            if (config.ForwardingFilter.Version != null)
+            {
+                cc.forwarding_filter.enable_version = true;
+                cc.forwarding_filter.version = config.ForwardingFilter.Version;
+            }
+            if (config.ForwardingFilter.Metadata != null)
+            {
+                cc.forwarding_filter.enable_metadata = true;
+                cc.forwarding_filter.metadata = config.ForwardingFilter.Metadata;
             }
         }
         cc.enable_use_hardware_encoder = config.UseHardwareEncoder != null;